refactor: move Equals string comparison method from Check.StringAssertions.cs to Check.Equals.cs
