+++ Code/Light.GuardClauses/Check.StringAssertions.cs
@@ -3,7 +3,6 @@ using System.Diagnostics.CodeAnalysis;
 using System.Text.RegularExpressions;
 using JetBrains.Annotations;
 using Light.GuardClauses.Exceptions;
-using Light.GuardClauses.FrameworkExtensions;
 using System.Runtime.CompilerServices;
 using NotNullAttribute = System.Diagnostics.CodeAnalysis.NotNullAttribute;
 
@@ -11,28 +10,6 @@ namespace Light.GuardClauses;
 
 public static partial class Check
 {
-    /// <summary>
-    /// Checks if the specified strings are equal, using the given comparison rules.
-    /// </summary>
-    /// <param name="string">The first string to compare.</param>
-    /// <param name="value">The second string to compare.</param>
-    /// <param name="comparisonType">One of the enumeration values that specifies the rules for the comparison.</param>
-    /// <returns>True if the two strings are considered equal, else false.</returns>
-    /// <exception cref="ArgumentException">Thrown when <paramref name="comparisonType"/> is no valid enum value.</exception>
-    [MethodImpl(MethodImplOptions.AggressiveInlining)]
-    public static bool Equals(this string? @string, string? value, StringComparisonType comparisonType)
-    {
-        if ((int) comparisonType < 6)
-            return string.Equals(@string, value, (StringComparison) comparisonType);
-        if (comparisonType == StringComparisonType.OrdinalIgnoreWhiteSpace)
-            return @string.EqualsOrdinalIgnoreWhiteSpace(value);
-        if (comparisonType == StringComparisonType.OrdinalIgnoreCaseIgnoreWhiteSpace)
-            return @string.EqualsOrdinalIgnoreCaseIgnoreWhiteSpace(value);
-
-        Throw.EnumValueNotDefined(comparisonType, nameof(comparisonType));
-        return false;
-    }
-
     /// <summary>
     /// Ensures that the string contains the specified substring, or otherwise throws a <see cref="SubstringException" />.
     /// </summary>
